Restructure project into agentdeal package
