+++ agentdeal/prompts/choose_cards_to_discard_prompt.j2
