+++ agentdeal/prompts/game_rules.j2
